pyproject.toml
[project]
name = "claude-memory-kit"
version = "0.2.0"
description = "Claude Memory Kit (CMK). Persistent memory for Claude. Embedded, zero-server."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Thierry Damiba"},
]
keywords = ["claude", "memory", "mcp", "llm", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp>=1.0",
    "qdrant-client>=1.12",
    "fastembed>=0.4",
    "httpx>=0.27",
    "click>=8.0",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "PyJWT[crypto]>=2.0",
]

[project.urls]
Homepage = "https://github.com/thierrypdamiba/claude-memory"
Repository = "https://github.com/thierrypdamiba/claude-memory"

[project.scripts]
cmk = "claude_memory_kit.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_memory_kit"]

[tool.hatch.build.targets.sdist]
include = ["/src/claude_memory_kit", "/pyproject.toml", "/README.md", "/LICENSE", "/.env.example"]
exclude = ["*.egg-info", ".env.local", "*.db", ".DS_Store"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests that are slow (fastembed model download)",
]

[tool.coverage.run]
source = ["claude_memory_kit"]
omit = ["src/claude_memory_kit/__main__.py"]

[tool.coverage.report]
fail_under = 90
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__",
    "log\\.",
]
